Create DOC.BOTS IoT TEMPRATURE DEVICE
this code reads the temprature and humidity from  dht11 and shows its on 20x4 lcd and uploaded it to thinkspeak and triggers an email if the temprature is increased and decreased from the defined limit. this code uploaded the readings on thingpseak with the interval of 20 seconds and email with the interval of 30minutes if the tempratures crosses the defined boundary.